refactor(numerical-semigroups): isolate presentation contracts
