+++ tests/math/numerical_semigroups/test_extended.py
@@ -32,10 +32,14 @@ from jacobian.math.numerical_semigroups._models import (
     ElementDeltaSetResult,
     ElementElasticityRequest,
     ElementElasticityResult,
+)
+from jacobian.math.numerical_semigroups._presentations import (
     MinimalPresentationRequest,
     MinimalPresentationResult,
     PresentationBinomialsRequest,
     PresentationBinomialsResult,
+    compute_minimal_presentation,
+    compute_presentation_binomials,
 )
 from jacobian.math.numerical_semigroups._operations import (
     compute_betti_elements,
@@ -45,8 +49,6 @@ from jacobian.math.numerical_semigroups._operations import (
     compute_element_catenary_degree,
     compute_element_delta_set,
     compute_element_elasticity,
-    compute_minimal_presentation,
-    compute_presentation_binomials,
 )
